Removed unnecessary data migrations, separated data migration to each classes
